+++ edison-status/src/main/java/de/otto/edison/status/controller/StatusController.java
@@ -1,7 +1,5 @@
 package de.otto.edison.status.controller;
 
-import de.otto.edison.status.domain.ApplicationStatus;
-import de.otto.edison.status.domain.StatusDetail;
 import de.otto.edison.status.indicator.ApplicationStatusAggregator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.context.properties.ConfigurationProperties;
@@ -9,13 +7,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.ModelAndView;
 
-import java.util.Collection;
-import java.util.LinkedHashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
-
-import static de.otto.edison.status.controller.ApplicationStatusRepresentation.statusRepresentationOf;
+import static de.otto.edison.status.controller.StatusRepresentation.statusRepresentationOf;
 import static java.time.ZonedDateTime.now;
 import static java.time.format.DateTimeFormatter.ofLocalizedDateTime;
 import static java.time.format.FormatStyle.LONG;
@@ -42,7 +34,7 @@ public class StatusController {
             produces = {"application/vnd.otto.monitoring.status+json", "application/json"},
             method = GET
     )
-    public ApplicationStatusRepresentation getStatusAsJson() {
+    public StatusRepresentation getStatusAsJson() {
         return statusRepresentationOf(aggregator.aggregatedStatus());
     }
 
@@ -52,20 +44,7 @@ public class StatusController {
             method = GET
     )
     public ModelAndView getStatusAsHtml() {
-        final ApplicationStatus applicationStatus = aggregator.aggregatedStatus();
         return new ModelAndView("status", "status", aggregator.aggregatedStatus());
-    }
-
-    private Collection<Map<String, ?>> statusDetails(final List<StatusDetail> statusDetails) {
-        return statusDetails.stream()
-                .map(detail ->
-                            new LinkedHashMap<String, Object>() {{
-                                put("key", detail.getName());
-                                put("status", detail.getStatus().name());
-                                put("message", detail.getMessage());
-                                put("additionalDetails", detail.getDetails());
-                            }})
-                .collect(Collectors.toList());
     }
 
 }