+++ edison-status/src/main/java/de/otto/edison/status/domain/ApplicationStatus.java
@@ -5,7 +5,10 @@ import net.jcip.annotations.Immutable;
 import java.util.ArrayList;
 import java.util.List;
 
+import static java.util.Collections.emptyList;
 import static java.util.Collections.unmodifiableList;
+import static java.util.Comparator.comparing;
+import static java.util.stream.Collectors.toList;
 
 @Immutable
 public final class ApplicationStatus {
@@ -13,13 +16,17 @@ public final class ApplicationStatus {
     public final ApplicationInfo application;
     public final SystemInfo system;
     public final VersionInfo vcs;
+    public final TeamInfo team;
     public final Status status;
     public final List<StatusDetail> statusDetails;
+    public final List<ServiceSpec> serviceSpecs;
 
     private ApplicationStatus(final ApplicationInfo application,
                               final SystemInfo system,
                               final VersionInfo vcs,
-                              final List<StatusDetail> details) {
+                              final TeamInfo team,
+                              final List<StatusDetail> details,
+                              final List<ServiceSpec> serviceSpecs) {
         this.status = details.stream()
                 .map(StatusDetail::getStatus)
                 .reduce(Status.OK, Status::plus);
@@ -27,13 +34,17 @@ public final class ApplicationStatus {
         this.application = application;
         this.system = system;
         this.vcs = vcs;
+        this.team = team;
+        this.serviceSpecs = serviceSpecs != null ? serviceSpecs.stream().sorted(comparing(spec->spec.appId)).collect(toList()) : emptyList();
     }
 
     public static ApplicationStatus applicationStatus(final ApplicationInfo applicationInfo,
                                                       final SystemInfo systemInfo,
                                                       final VersionInfo versionInfo,
-                                                      final List<StatusDetail> details) {
-        return new ApplicationStatus(applicationInfo, systemInfo, versionInfo, details);
+                                                      final TeamInfo teamInfo,
+                                                      final List<StatusDetail> details,
+                                                      final List<ServiceSpec> serviceSpecs) {
+        return new ApplicationStatus(applicationInfo, systemInfo, versionInfo, teamInfo, details, serviceSpecs);
     }
 
     @Override
@@ -46,9 +57,13 @@ public final class ApplicationStatus {
         if (application != null ? !application.equals(that.application) : that.application != null) return false;
         if (system != null ? !system.equals(that.system) : that.system != null) return false;
         if (vcs != null ? !vcs.equals(that.vcs) : that.vcs != null) return false;
+        if (team != null ? !team.equals(that.team) : that.team != null) return false;
         if (status != that.status) return false;
-        return !(statusDetails != null ? !statusDetails.equals(that.statusDetails) : that.statusDetails != null);
+        if (statusDetails != null ? !statusDetails.equals(that.statusDetails) : that.statusDetails != null)
+            return false;
+        if (serviceSpecs != null ? !serviceSpecs.equals(that.serviceSpecs) : that.serviceSpecs != null) return false;
 
+        return true;
     }
 
     @Override
@@ -56,8 +71,10 @@ public final class ApplicationStatus {
         int result = application != null ? application.hashCode() : 0;
         result = 31 * result + (system != null ? system.hashCode() : 0);
         result = 31 * result + (vcs != null ? vcs.hashCode() : 0);
+        result = 31 * result + (team != null ? team.hashCode() : 0);
         result = 31 * result + (status != null ? status.hashCode() : 0);
         result = 31 * result + (statusDetails != null ? statusDetails.hashCode() : 0);
+        result = 31 * result + (serviceSpecs != null ? serviceSpecs.hashCode() : 0);
         return result;
     }
 }