+++ edison-status/src/main/java/de/otto/edison/status/indicator/CachedApplicationStatusAggregator.java
@@ -2,9 +2,7 @@ package de.otto.edison.status.indicator;
 
 import de.otto.edison.status.domain.*;
 
-import java.util.Collections;
 import java.util.List;
-import java.util.Optional;
 
 import static de.otto.edison.status.domain.ApplicationStatus.applicationStatus;
 import static java.util.Collections.emptyList;
@@ -23,17 +21,23 @@ public class CachedApplicationStatusAggregator implements ApplicationStatusAggre
     private final ApplicationInfo applicationInfo;
     private final SystemInfo systemInfo;
     private final VersionInfo versionInfo;
+    private final TeamInfo teamInfo;
     private final List<StatusDetailIndicator> indicators;
+    private final List<ServiceSpec> serviceSpecs;
 
     public CachedApplicationStatusAggregator(final ApplicationInfo applicationInfo,
                                              final SystemInfo systemInfo,
                                              final VersionInfo versionInfo,
-                                             final List<StatusDetailIndicator> indicators) {
+                                             final TeamInfo teamInfo,
+                                             final List<StatusDetailIndicator> indicators,
+                                             final List<ServiceSpec> serviceSpecs) {
         this.applicationInfo = applicationInfo;
         this.systemInfo = systemInfo;
         this.versionInfo = versionInfo;
+        this.teamInfo = teamInfo;
         this.indicators = indicators;
-        this.cachedStatus = applicationStatus(applicationInfo, systemInfo, versionInfo, Collections.<StatusDetail>emptyList());
+        this.serviceSpecs = serviceSpecs;
+        this.cachedStatus = applicationStatus(applicationInfo, systemInfo, versionInfo, teamInfo, emptyList(), serviceSpecs);
     }
 
     @Override
@@ -43,11 +47,14 @@ public class CachedApplicationStatusAggregator implements ApplicationStatusAggre
 
     @Override
     public void update() {
-        final List<StatusDetail> allDetails = indicators
-                .stream()
-                .flatMap(i->i.statusDetails().stream())
-                .collect(toList());
-        this.cachedStatus = applicationStatus(applicationInfo, systemInfo, versionInfo, allDetails);
+        this.cachedStatus = applicationStatus(applicationInfo, systemInfo, versionInfo, teamInfo, getStatusDetails(indicators), serviceSpecs);
+    }
+
+    private static List<StatusDetail> getStatusDetails(final List<StatusDetailIndicator> indicators) {
+        return indicators
+                    .stream()
+                    .flatMap(i->i.statusDetails().stream())
+                    .collect(toList());
     }
 
 }