+++ example-status/src/main/java/de/otto/edison/example/status/ServiceSpecConfiguration.java
@@ -1,20 +1,20 @@
 package de.otto.edison.example.status;
 
-import de.otto.edison.about.spec.ServiceSpec;
+import de.otto.edison.status.domain.ServiceSpec;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-import static de.otto.edison.about.spec.Criticality.FUNCTIONAL_CRITICAL;
-import static de.otto.edison.about.spec.Expectations.lowExpectations;
-import static de.otto.edison.about.spec.ServiceSpec.serviceSpec;
-import static de.otto.edison.about.spec.ServiceType.TYPE_DATA_FEED;
-import static de.otto.edison.about.spec.ServiceType.serviceType;
+import static de.otto.edison.status.domain.Criticality.FUNCTIONAL_CRITICAL;
+import static de.otto.edison.status.domain.Expectations.lowExpectations;
+import static de.otto.edison.status.domain.ServiceSpec.serviceSpec;
+import static de.otto.edison.status.domain.ServiceType.TYPE_DATA_FEED;
+import static de.otto.edison.status.domain.ServiceType.serviceType;
 
 /**
  * An example for a configuration of some dependencies to other services.
  *
  * This information is gathered by Edison to provide information about such dependencies.
- * The data is exposed by the /internal/about API for external usage.
+ * The data is exposed by the /internal/status API for external usage.
  *
  * Created by guido on 07.01.16.
  */